enable mosaicml ui for more custom per module fsdp kwargs
